Fix long-running simulation stability issues
- Convert all metabolite quantities from Float to Double for improved
  numerical precision (prevents precision loss in ODE calculations with
  constants spanning 27+ orders of magnitude)
- Remove diagnostic println() calls from GlycolysisODE that were causing
  I/O bottleneck during long simulations (these ran hundreds of times
  per timestep inside computeDerivatives)
- Clean up redundant .toFloat() and .toDouble() conversion calls

Files modified:
- Metabolite.kt: amountMilliMoles Float → Double
- MetaboliteTransfer: percentage/actualAmount Float → Double
- QuantifiedMetabolite: percentageOfOutput Float → Double
- formatWeight(): Float → Double parameter
- All metabolite literal values: 0.0f → 0.0, 100.0f → 100.0, etc.

+++ composeApp/src/commonMain/kotlin/jon/sharp/metabolism/simulator/model/Metabolite.kt
@@ -42,6 +42,6 @@ enum class UpdateType {
 
 data class Metabolite(
     val type: MetaboliteType,
-    var amountMilliMoles: Float,
+    var amountMilliMoles: Double,
     val updateType: UpdateType = UpdateType.ACCUMULATE
 )
