+++ composeApp/src/jvmMain/kotlin/jon/sharp/metabolism/simulator/model/organ/blood/Blood.kt
@@ -15,17 +15,17 @@ class Blood: Organ(
     MetaboliteMap(
         Metabolite(
             MetaboliteType.GLUCOSE,
-            (5.0 * PhysicalConstants.Blood.TOTAL_LITERS).toFloat()
+            5.0 * PhysicalConstants.Blood.TOTAL_LITERS
         ),
         Metabolite(
             MetaboliteType.INSULIN,
             PhysicalConstants.Insulin.internationalUnitConcToMmol(
                 PhysicalConstants.Blood.FastingValues.basalInsulin
-            ).toFloat()
+            )
         ),
         Metabolite(
             MetaboliteType.GLUTAMIC_ACID,
-            (PhysicalConstants.Blood.FastingValues.basalBCAA * PhysicalConstants.Blood.TOTAL_LITERS).toFloat()
+            PhysicalConstants.Blood.FastingValues.basalBCAA * PhysicalConstants.Blood.TOTAL_LITERS
         )
     )
 )
